FLIZ-241 rename: EmptySearchResult 파일 위치 이동
